Fix producer timestamp docs
The producer timestamp is documented as being in microseconds but is actually expected to be milliseconds.  It is passed to `rd_kafka_producev` as-is and that function documents it as milliseconds.  Just to make sure I looked at the implementation of `rd_kafka_producev` and that puts the timestamp in the Kafka message unmodified.  The Kafka docs say that timestamp is in milliseconds.